refactor: primary key integer to long

+++ xcloud-iam-core/src/main/java/com/wl4g/iam/service/ContactGroupService.java
@@ -28,7 +28,7 @@ public interface ContactGroupService {
 
 	void save(ContactGroup contactGroup);
 
-	void del(Integer id);
+	void del(Long id);
 
 	List<ContactGroup> contactGroups(String name);
